requirements.txt
asgiref==3.7.2
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.7,<4.0
djangorestframework-simplejwt>=5.0,<6.0
psycopg2>=2.9,<3.0 
pillow>=8.3,<9.0
python-decouple==3.8
drf-yasg>=1.20,<2.0
gunicorn>=20.0,<21.0
pytz==2023.3.post1
sqlparse==0.4.4